requirements.txt
elasticsearch == 7.1.0
elasticsearch-dsl == 7.1.0
scrapy >= 1.8.0
bonobo == 0.6.4
nltk == 3.4.5
spacy == 2.2.3
fastapi == 0.52.0
uvicorn == 0.11.3
bcrypt == 3.1.7
pyjwt == 1.7.1
